+++ app/src/main/java/com/anthonyla/paperize/presentation/screens/sort/SortViewScreen.kt
@@ -332,7 +332,7 @@ fun SortViewScreen(
                                                     .clip(AppShapes.imageShape)
                                             )
                                         },
-                                        headlineContent = {
+                                        content = {
                                             Text(
                                                 text = currentFolder.name,
                                                 maxLines = 2,